awesome-ai-builds/advanced/multi-agent-news-analyst/requirements.txt
celery==5.4.0
crewai==0.95.0
openai==1.58.1
psycopg2-binary==2.9.10
pytest==8.3.4
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
google-generativeai==0.8.3
